removed all subscriptions from page.tsx and moved them to SetGameSubscriptions.tsx
